requirement.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic
python-multipart